Refactor where some of the code lives
