--- src/superintendent/distributed/__init__.py
@@ -1,1 +1,0 @@
-"""Interactive machine learning supervision using databases to distribute."""
