+++ src/superintendent/serialization.py
